+++ src/main/java/com/example/vehicleapi/controller/WorkflowController.java
@@ -27,7 +27,7 @@ public class WorkflowController {
 
     @GetMapping("/<built-in function id>")
     @Operation(summary = "Get by ID")
-    public WorkflowDto getById(@PathVariable Long id) {
+    public WorkflowDto getById(@PathVariable Integer id) {
         return mapper.toDto(repository.findById(id).orElseThrow());
     }
 
@@ -40,7 +40,7 @@ public class WorkflowController {
 
     @PutMapping("/<built-in function id>")
     @Operation(summary = "Update record")
-    public WorkflowDto update(@PathVariable Long id, @RequestBody WorkflowDto dto) {
+    public WorkflowDto update(@PathVariable Integer id, @RequestBody WorkflowDto dto) {
         Workflow entity = mapper.toEntity(dto);
         entity.setId(id);
         return mapper.toDto(repository.save(entity));
@@ -48,7 +48,7 @@ public class WorkflowController {
 
     @DeleteMapping("/<built-in function id>")
     @Operation(summary = "Delete record")
-    public void delete(@PathVariable Long id) {
+    public void delete(@PathVariable Integer id) {
         repository.deleteById(id);
     }
 }
