+++ src/main/java/com/example/vehicleapi/repository/AuctionFilterRepository.java
@@ -3,5 +3,5 @@ package com.example.vehicleapi.repository;
 import com.example.vehicleapi.model.AuctionFilter;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AuctionFilterRepository extends JpaRepository<AuctionFilter, Long> {
+public interface AuctionFilterRepository extends JpaRepository<AuctionFilter, Integer> {
 }
