+++ src/main/java/com/example/vehicleapi/repository/AuctionWatchRepository.java
@@ -3,5 +3,5 @@ package com.example.vehicleapi.repository;
 import com.example.vehicleapi.model.AuctionWatch;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AuctionWatchRepository extends JpaRepository<AuctionWatch, Long> {
+public interface AuctionWatchRepository extends JpaRepository<AuctionWatch, Integer> {
 }
