+++ src/main/java/com/example/vehicleapi/repository/ContactInfoRepository.java
@@ -3,5 +3,5 @@ package com.example.vehicleapi.repository;
 import com.example.vehicleapi.model.ContactInfo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ContactInfoRepository extends JpaRepository<ContactInfo, Long> {
+public interface ContactInfoRepository extends JpaRepository<ContactInfo, Integer> {
 }
