+++ src/main/java/com/example/vehicleapi/repository/VehicleConditionRepository.java
@@ -3,5 +3,5 @@ package com.example.vehicleapi.repository;
 import com.example.vehicleapi.model.VehicleCondition;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface VehicleConditionRepository extends JpaRepository<VehicleCondition, Long> {
+public interface VehicleConditionRepository extends JpaRepository<VehicleCondition, Integer> {
 }
