+++ src/main/java/com/example/vehicleapi/repository/VehicleConditionTypeRepository.java
@@ -3,5 +3,5 @@ package com.example.vehicleapi.repository;
 import com.example.vehicleapi.model.VehicleConditionType;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface VehicleConditionTypeRepository extends JpaRepository<VehicleConditionType, Long> {
+public interface VehicleConditionTypeRepository extends JpaRepository<VehicleConditionType, Integer> {
 }
