+++ src/main/java/com/example/vehicleapi/repository/VehicleMetadataRepository.java
@@ -3,5 +3,5 @@ package com.example.vehicleapi.repository;
 import com.example.vehicleapi.model.VehicleMetadata;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface VehicleMetadataRepository extends JpaRepository<VehicleMetadata, Long> {
+public interface VehicleMetadataRepository extends JpaRepository<VehicleMetadata, Integer> {
 }
