+++ src/main/java/com/example/vehicleapi/repository/VehicleStatusRepository.java
@@ -3,5 +3,5 @@ package com.example.vehicleapi.repository;
 import com.example.vehicleapi.model.VehicleStatus;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface VehicleStatusRepository extends JpaRepository<VehicleStatus, Long> {
+public interface VehicleStatusRepository extends JpaRepository<VehicleStatus, Integer> {
 }
