+++ src/main/java/com/example/vehicleapi/repository/VehicleStatusTypeRepository.java
@@ -3,5 +3,5 @@ package com.example.vehicleapi.repository;
 import com.example.vehicleapi.model.VehicleStatusType;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface VehicleStatusTypeRepository extends JpaRepository<VehicleStatusType, Long> {
+public interface VehicleStatusTypeRepository extends JpaRepository<VehicleStatusType, Integer> {
 }
