+++ src/main/java/com/example/vehicleapi/repository/WorkflowLogRepository.java
@@ -3,5 +3,5 @@ package com.example.vehicleapi.repository;
 import com.example.vehicleapi.model.WorkflowLog;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface WorkflowLogRepository extends JpaRepository<WorkflowLog, Long> {
+public interface WorkflowLogRepository extends JpaRepository<WorkflowLog, Integer> {
 }
