+++ src/main/java/com/example/vehicleapi/repository/WorkflowQueueRepository.java
@@ -3,5 +3,5 @@ package com.example.vehicleapi.repository;
 import com.example.vehicleapi.model.WorkflowQueue;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface WorkflowQueueRepository extends JpaRepository<WorkflowQueue, Long> {
+public interface WorkflowQueueRepository extends JpaRepository<WorkflowQueue, Integer> {
 }
