+++ src/main/java/com/example/vehicleapi/repository/WorkflowRepository.java
@@ -3,5 +3,5 @@ package com.example.vehicleapi.repository;
 import com.example.vehicleapi.model.Workflow;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface WorkflowRepository extends JpaRepository<Workflow, Long> {
+public interface WorkflowRepository extends JpaRepository<Workflow, Integer> {
 }
