+++ src/main/java/com/example/vehicleapi/repository/WorkflowStepRepository.java
@@ -3,5 +3,5 @@ package com.example.vehicleapi.repository;
 import com.example.vehicleapi.model.WorkflowStep;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface WorkflowStepRepository extends JpaRepository<WorkflowStep, Long> {
+public interface WorkflowStepRepository extends JpaRepository<WorkflowStep, Integer> {
 }
