+++ src/main/java/com/example/vehicleapi/repository/WorkflowStepTypeRepository.java
@@ -3,5 +3,5 @@ package com.example.vehicleapi.repository;
 import com.example.vehicleapi.model.WorkflowStepType;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface WorkflowStepTypeRepository extends JpaRepository<WorkflowStepType, Long> {
+public interface WorkflowStepTypeRepository extends JpaRepository<WorkflowStepType, Integer> {
 }
