+++ src/main/java/com/example/vehicleapi/repository/WorkflowTypeRepository.java
@@ -3,5 +3,5 @@ package com.example.vehicleapi.repository;
 import com.example.vehicleapi.model.WorkflowType;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface WorkflowTypeRepository extends JpaRepository<WorkflowType, Long> {
+public interface WorkflowTypeRepository extends JpaRepository<WorkflowType, Integer> {
 }
